addet to the coad to define a child

+++ src/Admin.java
@@ -1,6 +1,6 @@
 public class Admin extends Staff {
 
-    public Admin (MainSystem main, String fName, String lName, int phoneNumber, int salary, String position){
+    public Admin (MainSystem main, String fName, String lName, long phoneNumber, int salary, String position){
         super(main, fName, lName, phoneNumber, salary, position);
     }
