+++ src/Staff.java
@@ -3,7 +3,7 @@ public class Staff extends People {
     int salary;
     String position;
 
-    public Staff(MainSystem main, String fName, String lName, int phoneNumber, int salary, String position){
+    public Staff(MainSystem main, String fName, String lName, long phoneNumber, int salary, String position){
         super(main, fName, lName, phoneNumber);
 
         setSalary(salary);
